feat: refactor audio player component and implement track fetching logic
